Fixed User Sign Up
