+++ frontend/src/app/authentication/Callback.tsx
@@ -1,0 +1,5 @@
+import { AuthenticateWithRedirectCallback } from "@clerk/clerk-react";
+
+export default function Callback () {
+  return <AuthenticateWithRedirectCallback />;
+};
